Move all code to OleDbStruct.cs
