+++ src/libraries/System.Data.OleDb/src/OleDbStruct.cs
@@ -307,9 +307,22 @@ namespace System.Data.OleDb
     }
 #endif
 
-    [StructLayout(LayoutKind.Sequential, Pack = 2)]
-    internal sealed partial class tagDBPROP_x86
+    internal interface ItagDBPROP
     {
+        OleDbPropertyStatus dwStatus { get; }
+        object vValue { get; }
+        int dwPropertyID { get; }
+    }
+
+    [StructLayout(LayoutKind.Sequential, Pack = 2)]
+    internal sealed class tagDBPROP_x86 : ItagDBPROP
+    {
+        OleDbPropertyStatus ItagDBPROP.dwStatus => this.dwStatus;
+
+        object ItagDBPROP.vValue => this.vValue;
+
+        int ItagDBPROP.dwPropertyID => this.dwPropertyID;
+
         internal int dwPropertyID;
         internal int dwOptions;
         internal OleDbPropertyStatus dwStatus;
@@ -332,8 +345,13 @@ namespace System.Data.OleDb
     }
 
     [StructLayout(LayoutKind.Sequential, Pack = 8)]
-    internal sealed partial class tagDBPROP
+    internal sealed class tagDBPROP : ItagDBPROP
     {
+        OleDbPropertyStatus ItagDBPROP.dwStatus => this.dwStatus;
+
+        object ItagDBPROP.vValue => this.vValue;
+
+        int ItagDBPROP.dwPropertyID => this.dwPropertyID;
         internal int dwPropertyID;
         internal int dwOptions;
         internal OleDbPropertyStatus dwStatus;
@@ -471,9 +489,28 @@ namespace System.Data.OleDb
     }
 #endif
 
-    [StructLayout(LayoutKind.Sequential, Pack = 2)]
-    internal sealed partial class tagDBPROPINFO_x86
+    internal interface ItagDBPROPINFO
     {
+        int dwPropertyID { get; }
+        int dwFlags { get; }
+        int vtType { get; }
+        object vValue { get; }
+        string pwszDescription { get; }
+    }
+
+    [StructLayout(LayoutKind.Sequential, Pack = 2)]
+    internal sealed class tagDBPROPINFO_x86 : ItagDBPROPINFO
+    {
+        int ItagDBPROPINFO.dwPropertyID => this.dwPropertyID;
+
+        int ItagDBPROPINFO.dwFlags => this.dwFlags;
+
+        int ItagDBPROPINFO.vtType => this.vtType;
+
+        object ItagDBPROPINFO.vValue => this.vValue;
+
+        string ItagDBPROPINFO.pwszDescription => this.pwszDescription;
+
         [MarshalAs(UnmanagedType.LPWStr)] internal string pwszDescription;
 
         internal int dwPropertyID;
@@ -489,8 +526,17 @@ namespace System.Data.OleDb
     }
 
     [StructLayout(LayoutKind.Sequential, Pack = 8)]
-    internal sealed partial class tagDBPROPINFO
+    internal sealed class tagDBPROPINFO : ItagDBPROPINFO
     {
+        int ItagDBPROPINFO.dwPropertyID => this.dwPropertyID;
+
+        int ItagDBPROPINFO.dwFlags => this.dwFlags;
+
+        int ItagDBPROPINFO.vtType => this.vtType;
+
+        object ItagDBPROPINFO.vValue => this.vValue;
+
+        string ItagDBPROPINFO.pwszDescription => this.pwszDescription;
         [MarshalAs(UnmanagedType.LPWStr)] internal string pwszDescription;
 
         internal int dwPropertyID;
@@ -522,5 +568,44 @@ namespace System.Data.OleDb
         internal IntPtr rgPropertyIDs;
         internal int cPropertyIDs;
         internal Guid guidPropertySet;
+    }
+
+    internal static class ArchitectureSpecificHelpers
+    {
+        internal static ItagDBPROPINFO CreateTagDbPropInfo()
+        {
+            if (ODB.IsRunningOnX86)
+            {
+                return new tagDBPROPINFO_x86();
+            }
+            else
+            {
+                return new tagDBPROPINFO();
+            }
+        }
+
+        internal static ItagDBPROP CreateTagDbProp(int propertyID, bool required, object value)
+        {
+            if (ODB.IsRunningOnX86)
+            {
+                return new tagDBPROP_x86(propertyID, required, value);
+            }
+            else
+            {
+                return new tagDBPROP(propertyID, required, value);
+            }
+        }
+
+        internal static ItagDBPROP CreateTagDbProp()
+        {
+            if (ODB.IsRunningOnX86)
+            {
+                return new tagDBPROP_x86();
+            }
+            else
+            {
+                return new tagDBPROP();
+            }
+        }
     }
 }
