+++ src/libraries/System.Data.OleDb/src/System.Data.OleDb.csproj
@@ -17,7 +17,6 @@
       <Link>Common\System\Data\Common\MultipartIdentifier.cs</Link>
     </Compile>
     <Compile Include="AdapterSwitches.cs" />
-    <Compile Include="ArchitectureSpecificHelpers.cs" />
     <Compile Include="ColumnBinding.cs" />
     <Compile Include="DbBindings.cs" />
     <Compile Include="DbConnectionOptions.cs" />
